requirements.txt
pycryptodome==3.21.0
bcrypt==4.3.0
python-dotenv